Added Conexant ADSL PCI support:
New kernel module from a common source tarball

Userspace bits from the common source tarball have the prog to load the
firmware and do the sync.

Minor tweaks to the dial scripts and setup program. The same pppd options are
used as the bewan.

Messy bit is the cnxadsl.conf file, which is .in'd into the coretree. A script,
writecnxadslconfig.pl includes the VCI/VPI settings at the bottom of it.


git-svn-id: https://smoothwall.svn.sourceforge.net/svnroot/smoothwall/trunk@676 2e32aa7d-f303-0410-ac85-8a20d3371fc6