Fixes for building smoothie on the very latest distros:
dash patches to make tcl and glibc build when /bin/sh links to dash, as it
does in ubuntu.  Also glibc uses /bin/echo instead of the shell builtin
because dash echo is crappy.  prebuild2 makes a symlink for it.

Bump binutils so it can uderstand the hosts libc when building gcc.

Reordered coreutils so /usr/bin/install is available earlier.


git-svn-id: https://smoothwall.svn.sourceforge.net/svnroot/smoothwall/trunk@805 2e32aa7d-f303-0410-ac85-8a20d3371fc6