Fixed validation of in-use ports on portforward page.
It will now only flag as "in use" if the external address is the same too,
so you can forward from multiple outside IPs.

Also tweaked the handling of the two port fields on error, so the port
value is retained and not lost, when a data entry error occures.


git-svn-id: https://smoothwall.svn.sourceforge.net/svnroot/smoothwall/trunk@874 2e32aa7d-f303-0410-ac85-8a20d3371fc6