Replaced Apache 1.3 with Apache 2.2.
Apache 1.3 is waay past end of life. No fixes have been made in 2 years.

Apache now sits in /usr/apache2/sbin/httpd. The config remains in /etc/httpd.

The biggest change is to rework the config file, but the functionality is
the same more or less. We now have mod_deflate, which could be useful for
accessing the smoothie remotely.

The changepw.cgi and setup changes are because htpasswd has moved from
/usr/bin to /usr/apache2/sbin.


git-svn-id: https://smoothwall.svn.sourceforge.net/svnroot/smoothwall/trunk@1054 2e32aa7d-f303-0410-ac85-8a20d3371fc6