Work around a problem with /sbin/hotplug segfaults at boot.
An issue with linux kernel is the /sbin/hotplug method for loading firmware
(used for NIC firmware loading in smoothie as loading NICs is no done with udev)
can sometimes cause segfaults to be reported because /sbin/hotplug is being
run before we start /init.  This seems to be a long running issue:

http://lkml.indiana.edu/hypermail/linux/kernel/0701.1/1966.html

The fix, from that guy's post, is to bring /sbin/hotplug into exisitance in
the /init script so the kernel can't start it before it would segfault.