Fix some dodgy paramter handling.
iptools.cgi had a longstanding issue whereby it would not show ANY errors.
But when this is fixed, it reveals an injection attack whereby the input is
emitted into an errorbox unchanged. So add some parsing to make sure
hostnames have only DNS characters.

Also add some sanity checking the query string, as the help.cgi sees it. This
avoids the help.cgi potentially being used to read arbitary files.