Revert the bump to tar, for now.
This is because the newer tar has different behavior when it comes to
extracting an archive over a path which contains a symlink. New tar will
replace the symlink with a directory, vs old tar which follows the link and
puts the files at that path.  This behavior is needed, at least for 64bit
(lib and lib64), and probably other places.

There's an override for this behavior, but until we do a full audit this
change is too risky.