Handle changed NICs more smoothly.
But there is a problem. When the runtime init runs udev, udev can
add a 'new' NIC with the same name as an existing NIC. This causes
early boot to pause for about a minute. Other than this pause, the NIC
change process works quite well. This problem may be limited to VMs
that use a single driver for all NICs; it might not be a problem on
systems that use real NICs. This should be solved, but since it only
happens when NICs are added 'in the middle of' existing NICs, I'll
let it slide for now: NICs aren't added too often.