Modernize clock handling.
Re-do the UI. Default to pool.ntp.org (their DNS is geo-aware). Allow manual, periodic or
automatic operation.

Tweak gettime.pl to run only if time is enabled and the method is Periodic.

Tweak writentpd.pl to use the single available server (NTP_SERVER).

Replace the 'time' entries in en.pl with all new entries.

Tweak the smoothd plugin to start ntpd only if the method is 'Automatic' and time
regulation is enabled.