Pull in Chris' UI/script work to-date, with some tiny changes.
Add a comment to advstatus.cgi noting that '$3' is correct due to nested groups.
Tweak bandwidthbars.cgi to make it clear that $iface and $value are local to the loop
to guarantee that they are fresh each time through the loop.
Clear patches/available; this file is populated fresh each time the system checks for
updates. Only the installed file is pre-populated with the updates already applied in
the service pack.