Finally, udev-handling code that works on KVM and the old Y2k PIII. Both the basic and advanced installers seem to install reliably.
The trick to all this is figuring out when to reload the rules, when to trigger
a device change, when to wait pause a moment, when to wait for udev to settle,
when to generate a general trigger of all devices, ad infinitum. Udev is a little
unpredictable in some respects.